+++ payment-element-test-pages/src/main/java/com/stripe/paymentelementtestpages/FormPage.kt
@@ -59,7 +59,7 @@ class FormPage(
         composeTestRule.waitUntil {
             composeTestRule
                 .onAllNodes(hasTestTag(FORM_ELEMENT_TEST_TAG))
-                .fetchSemanticsNodes()
+                .fetchSemanticsNodes(atLeastOneRootRequired = false)
                 .isNotEmpty()
         }
     }