+++ payment-element-test-pages/src/main/java/com/stripe/paymentelementtestpages/PaymentElementPageConstants.kt
@@ -1,0 +1,3 @@
+package com.stripe.paymentelementtestpages
+
+const val DEFAULT_PE_PAGE_UI_TIMEOUT = 5000L