+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentOptionsViewModel.kt
@@ -40,6 +40,7 @@ import com.stripe.android.paymentsheet.model.PaymentSelection.Link
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen.AddFirstPaymentMethod
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen.SelectSavedPaymentMethods
+import com.stripe.android.paymentsheet.repositories.PaymentMethodMessagePromotionsHelper
 import com.stripe.android.paymentsheet.repositories.SavedPaymentMethodRepository
 import com.stripe.android.paymentsheet.state.WalletsProcessingState
 import com.stripe.android.paymentsheet.state.WalletsState
@@ -78,6 +79,7 @@ internal class PaymentOptionsViewModel @Inject constructor(
     mode: EventReporter.Mode,
     customerStateHolderFactory: CustomerStateHolder.Factory,
     @ViewModelScope customViewModelScope: CoroutineScope,
+    private val paymentMethodMessagePromotionsHelper: PaymentMethodMessagePromotionsHelper
 ) : BaseSheetViewModel(
     config = args.configuration,
     eventReporter = eventReporter,
@@ -437,7 +439,7 @@ internal class PaymentOptionsViewModel @Inject constructor(
                 viewModel = this,
                 paymentMethodMetadata = paymentMethodMetadata,
                 customerStateHolder = customerStateHolder,
-                paymentMethodMessagePromotionsHelper = null
+                paymentMethodMessagePromotionsHelper = paymentMethodMessagePromotionsHelper
             )
         }
         val target = if (args.state.showSavedPaymentMethods) {
@@ -452,7 +454,7 @@ internal class PaymentOptionsViewModel @Inject constructor(
             val interactor = DefaultAddPaymentMethodInteractor.create(
                 viewModel = this,
                 paymentMethodMetadata = paymentMethodMetadata,
-                paymentMethodMessagePromotionsHelper = null
+                paymentMethodMessagePromotionsHelper = paymentMethodMessagePromotionsHelper
             )
             AddFirstPaymentMethod(interactor = interactor)
         }
@@ -467,7 +469,7 @@ internal class PaymentOptionsViewModel @Inject constructor(
                 val interactor = DefaultAddPaymentMethodInteractor.create(
                     viewModel = this@PaymentOptionsViewModel,
                     paymentMethodMetadata = paymentMethodMetadata,
-                    paymentMethodMessagePromotionsHelper = null
+                    paymentMethodMessagePromotionsHelper = paymentMethodMessagePromotionsHelper
                 )
                 add(
                     PaymentSheetScreen.AddAnotherPaymentMethod(interactor = interactor)