+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/analytics/DefaultEventReporter.kt
@@ -419,9 +419,9 @@ internal class DefaultEventReporter @Inject internal constructor(
         fireEvent(PaymentSheetEvent.TapToAdd.Started(mode))
     }
 
-    override fun onCardAddedWithTapToAdd() {
+    override fun onCardAddedWithTapToAdd(canCollectLinkInput: Boolean) {
         val duration = durationProvider.end(DurationProvider.Key.TapToAdd)
-        fireEvent(PaymentSheetEvent.TapToAdd.CardAdded(mode, duration))
+        fireEvent(PaymentSheetEvent.TapToAdd.CardAdded(mode, duration, canCollectLinkInput))
     }
 
     override fun onTapToAddCanceled(source: EventReporter.TapToAddCancelSource) {
@@ -429,12 +429,12 @@ internal class DefaultEventReporter @Inject internal constructor(
         fireEvent(PaymentSheetEvent.TapToAdd.Canceled(mode, source, duration))
     }
 
-    override fun onTapToAddContinueAfterCardAdded() {
-        fireEvent(PaymentSheetEvent.TapToAdd.ContinueAfterCardAdded(mode))
+    override fun onTapToAddContinueAfterCardAdded(completedLinkInput: Boolean?) {
+        fireEvent(PaymentSheetEvent.TapToAdd.ContinueAfterCardAdded(mode, completedLinkInput))
     }
 
-    override fun onTapToAddConfirm() {
-        fireEvent(PaymentSheetEvent.TapToAdd.Confirm(mode))
+    override fun onTapToAddConfirm(recollectedCvc: Boolean) {
+        fireEvent(PaymentSheetEvent.TapToAdd.Confirm(mode, recollectedCvc))
     }
 
     override fun onFailedToAddCardWithTapToAdd(message: String) {