+++ paymentsheet/src/test/java/com/stripe/android/common/taptoadd/ui/DefaultTapToAddCardAddedInteractorTest.kt
@@ -30,44 +30,51 @@ import com.stripe.android.ui.core.R as StripeUiCoreR
 
 internal class DefaultTapToAddCardAddedInteractorTest {
     @Test
-    fun `initial state is correct`() = runScenario(
+    fun `Continue mode with Link disabled`() = runScenario(
         paymentMethod = PaymentMethodFactory.card().update(
             last4 = "1234",
             brand = CardBrand.MasterCard,
             addCbcNetworks = false,
         ),
-        initialLinkState = SavedPaymentMethodLinkFormHelper.State.Unused,
+        tapToAddMode = TapToAddMode.Continue,
+        linkElement = null,
     ) {
         val state = interactor.state.value
         assertThat(state.cardBrand).isEqualTo(CardBrand.MasterCard)
         assertThat(state.last4).isEqualTo("1234")
         assertThat(state.title).isEqualTo(R.string.stripe_tap_to_add_card_added_title.resolvableString)
-        assertThat(state.primaryButton.label)
-            .isEqualTo(StripeUiCoreR.string.stripe_continue_button_label.resolvableString)
-        assertThat(state.primaryButton.enabled).isTrue()
-        assertThat(state.form.elements).containsExactly(fakeLinkFormHelper.formElement)
+        assertThat(state.primaryButton).isNull()
+        assertThat(state.form.elements).isEmpty()
         assertThat(state.form.enabled).isTrue()
     }
 
     @Test
-    fun `state primary button is disabled when link helper is Incomplete`() = runScenario(
-        initialLinkState = SavedPaymentMethodLinkFormHelper.State.Complete(mockUserInput()),
+    fun `Complete mode with unused link shows primary button`() = runScenario(
+        paymentMethod = PaymentMethodFactory.card(last4 = "4242"),
+        tapToAddMode = TapToAddMode.Complete,
     ) {
-        interactor.state.test {
-            assertThat(awaitItem().primaryButton.enabled).isTrue()
-            fakeLinkFormHelper.updateState(SavedPaymentMethodLinkFormHelper.State.Incomplete)
-            assertThat(awaitItem().primaryButton.enabled).isFalse()
-        }
+        val state = interactor.state.value
+        assertThat(state.primaryButton).isNotNull()
+        assertThat(state.primaryButton?.label)
+            .isEqualTo(StripeUiCoreR.string.stripe_continue_button_label.resolvableString)
+        assertThat(state.primaryButton?.enabled).isTrue()
     }
 
     @Test
-    fun `state primary button is enabled when link helper emits Unused`() = runScenario(
-        initialLinkState = SavedPaymentMethodLinkFormHelper.State.Incomplete,
+    fun `state primary button is disabled when link available and helper is Incomplete`() = runScenario(
+        tapToAddMode = TapToAddMode.Complete,
+        initialLinkState = SavedPaymentMethodLinkFormHelper.State.Complete(mockUserInput()),
     ) {
         interactor.state.test {
-            assertThat(awaitItem().primaryButton.enabled).isFalse()
-            fakeLinkFormHelper.updateState(SavedPaymentMethodLinkFormHelper.State.Unused)
-            assertThat(awaitItem().primaryButton.enabled).isTrue()
+            val initialPrimaryButton = awaitItem().primaryButton
+            assertThat(initialPrimaryButton).isNotNull()
+            assertThat(initialPrimaryButton?.enabled).isTrue()
+
+            fakeLinkFormHelper.updateState(SavedPaymentMethodLinkFormHelper.State.Incomplete)
+
+            val updatedPrimaryButton = awaitItem().primaryButton
+            assertThat(updatedPrimaryButton).isNotNull()
+            assertThat(updatedPrimaryButton?.enabled).isFalse()
         }
     }
 
@@ -80,14 +87,26 @@ internal class DefaultTapToAddCardAddedInteractorTest {
     }
 
     @Test
-    fun `PrimaryButtonPressed in Continue mode calls onContinue with payment method`() =
+    fun `PrimaryButtonPressed reports completed link input false when link form is available but unused`() =
         runScenario(
-            tapToAddMode = TapToAddMode.Continue,
-            initialLinkState = SavedPaymentMethodLinkFormHelper.State.Unused,
+            tapToAddMode = TapToAddMode.Complete,
+            initialLinkState = SavedPaymentMethodLinkFormHelper.State.Unused
         ) {
             interactor.performAction(TapToAddCardAddedInteractor.Action.PrimaryButtonPressed)
 
-            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isNotNull()
+            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isFalse()
+            onConfirmCalls.awaitItem()
+        }
+
+    @Test
+    fun `PrimaryButtonPressed in Continue mode calls onContinue with payment method`() =
+        runScenario(
+            tapToAddMode = TapToAddMode.Continue,
+            linkElement = null,
+        ) {
+            interactor.performAction(TapToAddCardAddedInteractor.Action.PrimaryButtonPressed)
+
+            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isNull()
             val selection = onContinueCalls.awaitItem()
             assertThat(selection.paymentMethod).isEqualTo(paymentMethod)
             assertThat(selection.linkInput).isNull()
@@ -101,7 +120,7 @@ internal class DefaultTapToAddCardAddedInteractorTest {
         ) {
             interactor.performAction(TapToAddCardAddedInteractor.Action.PrimaryButtonPressed)
 
-            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isNotNull()
+            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isTrue()
             val selection = onContinueCalls.awaitItem()
             assertThat(selection.paymentMethod).isEqualTo(paymentMethod)
             assertThat(selection.linkInput).isEqualTo(mockUserInput())
@@ -111,22 +130,46 @@ internal class DefaultTapToAddCardAddedInteractorTest {
     fun `PrimaryButtonPressed in Complete mode calls onConfirm with payment method`() =
         runScenario(
             tapToAddMode = TapToAddMode.Complete,
-            initialLinkState = SavedPaymentMethodLinkFormHelper.State.Unused,
+            linkElement = null,
         ) {
             interactor.performAction(TapToAddCardAddedInteractor.Action.PrimaryButtonPressed)
 
-            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isNotNull()
+            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isNull()
             val (confirmedPm, linkInput) = onConfirmCalls.awaitItem()
             assertThat(confirmedPm).isEqualTo(paymentMethod)
             assertThat(linkInput).isNull()
         }
 
     @Test
+    fun `ScreenShown continues when primary button is hidden in Continue mode & no link elements`() = runScenario(
+        tapToAddMode = TapToAddMode.Continue,
+        linkElement = null,
+    ) {
+        interactor.performAction(TapToAddCardAddedInteractor.Action.ScreenShown)
+
+        val selection = onContinueCalls.awaitItem()
+
+        assertThat(selection.paymentMethod).isEqualTo(paymentMethod)
+        assertThat(selection.linkInput).isNull()
+    }
+
+    @Test
+    fun `ScreenShown does not continue when primary button is shown`() = runScenario(
+        tapToAddMode = TapToAddMode.Complete,
+        initialLinkState = SavedPaymentMethodLinkFormHelper.State.Unused,
+    ) {
+        interactor.performAction(TapToAddCardAddedInteractor.Action.ScreenShown)
+
+        onContinueCalls.expectNoEvents()
+        onConfirmCalls.expectNoEvents()
+    }
+
+    @Test
     fun `close should stop any more updates from link form helper from being received`() = runScenario(
         initialLinkState = SavedPaymentMethodLinkFormHelper.State.Unused,
     ) {
         interactor.state.test {
-            assertThat(awaitItem().primaryButton.enabled).isTrue()
+            assertThat(awaitItem().primaryButton).isNotNull()
             interactor.close()
             fakeLinkFormHelper.updateState(SavedPaymentMethodLinkFormHelper.State.Incomplete)
             expectNoEvents()
@@ -141,7 +184,7 @@ internal class DefaultTapToAddCardAddedInteractorTest {
         ) {
             interactor.performAction(TapToAddCardAddedInteractor.Action.PrimaryButtonPressed)
 
-            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isNotNull()
+            assertThat(fakeEventReporter.tapToAddContinueAfterCardAddedCalls.awaitItem()).isTrue()
             val (confirmedPaymentMethod, confirmedLinkInput) = onConfirmCalls.awaitItem()
             assertThat(confirmedPaymentMethod).isEqualTo(paymentMethod)
             assertThat(confirmedLinkInput).isEqualTo(mockUserInput())
@@ -153,6 +196,7 @@ internal class DefaultTapToAddCardAddedInteractorTest {
         paymentMethod: PaymentMethod = PaymentMethodFactory.card(last4 = "4242"),
         tapToAddMode: TapToAddMode = TapToAddMode.Continue,
         initialLinkState: SavedPaymentMethodLinkFormHelper.State = SavedPaymentMethodLinkFormHelper.State.Unused,
+        linkElement: FormElement? = FakeFormElement,
         block: suspend Scenario.() -> Unit,
     ) = runTest {
         val onContinueCalls = Turbine<PaymentSelection.Saved>()
@@ -160,7 +204,7 @@ internal class DefaultTapToAddCardAddedInteractorTest {
         val fakeEventReporter = FakeEventReporter()
         val fakeLinkFormHelper = FakeSavedPaymentMethodLinkFormHelper(
             initialState = initialLinkState,
-            formElement = FakeFormElement,
+            formElement = linkElement,
         )
 
         val scenario = Scenario(
@@ -197,7 +241,7 @@ internal class DefaultTapToAddCardAddedInteractorTest {
 
     private class FakeSavedPaymentMethodLinkFormHelper(
         initialState: SavedPaymentMethodLinkFormHelper.State,
-        override val formElement: FormElement,
+        override val formElement: FormElement?,
     ) : SavedPaymentMethodLinkFormHelper {
         private val _state = MutableStateFlow(initialState)