+++ tap-to-add-testing/src/main/java/com/stripe/android/tta/testing/TapToAddHelpers.kt
@@ -1,0 +1,21 @@
+package com.stripe.android.tta.testing
+
+import androidx.compose.ui.test.junit4.ComposeTestRule
+import androidx.compose.ui.test.onAllNodesWithTag
+import com.stripe.android.common.taptoadd.ui.TAP_TO_ADD_LAYOUT_TEST_TAG
+
+internal fun ComposeTestRule.waitUntilLayoutWithPrimaryButtonMissing(
+    buttonTestTag: String,
+) {
+    waitUntil(DEFAULT_UI_TIMEOUT) {
+        onAllNodesWithTag(buttonTestTag)
+            .fetchSemanticsNodes(atLeastOneRootRequired = false)
+            .isEmpty()
+    }
+
+    waitUntil(DEFAULT_UI_TIMEOUT) {
+        onAllNodesWithTag(TAP_TO_ADD_LAYOUT_TEST_TAG)
+            .fetchSemanticsNodes(atLeastOneRootRequired = false)
+            .isEmpty()
+    }
+}