+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 
+NEXT_VERSION_BUMP: PATCH
+
 ## XX.XX.XX - 20XX-XX-XX
+
+## 23.4.0 - 2026-04-14
+### Crypto Onramp (Private Preview)
+* [ADDED][12740](https://github.com/stripe/stripe-android/pull/12740) Ability to Receive L0 KYC Info From Google Pay
 
 ### Identity
 * [CHANGED][12811](https://github.com/stripe/stripe-android/pull/12811) Updated the Identity example app to use the new example API endpoint.